+++ hata/main/commands/default/__init__.py
@@ -1,5 +1,7 @@
 __all__ = ()
 
+from .scaffold import *
+
 from .help import *
 from .interpreter import *
 from .run import *